refactor(perl-lexer): split lexer config and limits into SRP modules (#0000)
